unlink: change how deleted entries are in memory
Previously, when removing an entry from a directory, the name was just
set to zero and size aswell. This was easy because without cache, after
updating memory item, it was serialized and this step was ignoring empty
names.

With the cache introduced, it's possible to get directory contents in
cache and thus having empty name and size in directory list, this cause
readdir unconsistancy and other problems.

This new implementation just swap last entry of a directory with the
directory to remove and shrink directory size to drop last element. This
last element is freed in the meantime.

Doing this broke unlink which was using old reference still in memory,
this was fixed aswell by keeping target inode as copy before removing
entry.