cache: implement block cache
Implement a block cache layer. Cache system is not implemented in
« filesystem performance » wise but more « backend optimization ».

It's way more important that 0-db size doesn't blow up behind than
getting lot of performance. This current cache implementation prevent
0-db to blow by sending blocks too quickly and rewrite them too often.

Current implementation doesn't use cache for read, unless a block is set
to be written, nothing is in cache. As soon as inode is in cache, blocks
written are in cache aswell and are flushed on file close.

In addition, a special hackish way to support memory was used, the ino
field on inode struct is used to keep track if inode is in cache or not
to know if object can be freed or not automatically, this fix the major
memory leak introduced by cache.

This implementation fix read aswell, now reading file with hole works.
Holes are well supported in the backend, 5 bytes written in a file with
10 MB hole will use 5 bytes in the backend, not 5 MB + 5 bytes.

Current implementation support the Linux Kernel build successfully over
fuse and use no noticable overhead in the backend.