cache: support block limitation and offloading
Previous block caching system were fine and quite good, but it was not
limited and could eat all the memory, it was a first implementation.

Previously, as soon as a file were opened, all write made inside that
file were cached in memory and were flushed to the backend only on file
close. This were ideal to reduce amount of write in the backend but if
you want to write a 8GB file, you'll end up with 8GB memory used.

This commit introduce a limited blocksize kept in memory, without
sending any blocks to data namespace. This feature use the extra
temporary namespace added previously. This temporary namespace needs to
be non public and password protected, which enable 'FLUSH' command to
works on it and discard the namespace content. Since all data in that
namespace are intended to be temporary, we don't care about history,
etc. it's there just for temporary storage.

Note: the FLUSH request is not yet supported on zdbfs.

There is couple changes here:
 - Last access time for inode is now a double with timestamp stored with
   microseconds.
 - There is an access time tracking per block aswell, using same double.
 - We keep all written blocks in memory but not all the data, entry have
   an 'online' flag to know if data is available or not

When a block write request is initiated, the first step is checking for
cache. There is a limited amount of blocks which can stay in memory (aka
online). If too many blocks are online, the zdbfs_cache_block_delegate
function try to find a block to offload (delegate). For now, this block
is just the olest block (old by access time), this block is sent to the
temporary namespace and flagged offline. Data buffer is freed.

Since all blocks written entries are kept in memory, if that block cache
request is made again, the block is fetched from temporary namespace and
restored in memory, so it can be updated.

When closing a file, each blocks are reached to be flushed and if data
were not available at that time, it's fetched first before being sent to
data namespace.

When this is done, temporary namespace should be checked to see if FLUSH
is possible (aka no more useful data remains).

Note: obviously, this introduce time overhead when offloading data and
fetching data back when needed, how often this will be done need to be
confugured by max block cache size.