cache: avoid temporary cache on linear write
Temporary cache is really useful and works great when block needs
caching for read-write update but is innefficient when doing regular
linear file write, like file copy or continuous write.

This commit implement a try-to-guess if current file write is linear or
random, if all cached blocks are aligned and ordered, blocksize full and
time incremented, we assume this is a regular linear write. On that
case, write block directly to datablock namespace and avoid temporary
namespace.

If any previous condition doesn't match, we use temporary cache.