ioctl: implement basic ioctl with snapshot request
Implement basic ioctl handler, with a single handler for now:
  snapshot request

A snapshot is just a timestamp for the database, this timestamp is
exactly when you want to rollback, this works for any point of time in
the database. Layout will change later to add more precision on the
timestamp.

This commit introduce ioctl handler for later usage and add a first
ioctl request 'ZDBFS_IOCTL_SNAPSHOT' which tell the filesystem we would
like a snapshot now, we need to flush cache to ensure database is fully
populated with current state, in order to avoid inconsistancy with
timestamp and real data in backend.

Return value of ioctl is the timestamp which can be used to rollback.