zdb: keep namespace per redis context
Add current selected namespace to redis context private data, which is
used now in debug/error message to specify in which namespace the error
is related.

Previously it was quite difficult to identify an error, eg: error
"Key not found" returned was not enough to identify which context and
which namespace was involved, now yes.