rmdir: keep cache sync and delete inode from backend
Call 'rmdir' was not removing target inode from the backend, only
updating the parent inode to remove the link, but target inode was still
physically present in the backend. This doesn't lead to any error but
keep unneeded inode.

The cache is also cleared to avoid flushing non-available inode anymore.