zdb: recover from error and improve code segmentation
Introduce a redis error recoverer. In case of zdb/redis error,
redisContext object should not be used anymore, as soon as the
connection is closed, it cannot be used anymore.

There is a check on each call now, which detect that response is
critical and a reconnection is automatically performed, trying to keep
connection state always sane.

In addition, this ensure that any redisContext is cleanup after an error
and doesn't keep growing it's internal buffer by submitting more request
later. This caused an always-memory-growing (not leak) issue.

When submitting a payload larger than 8 MB, zdb close the connection to
avoid any too large request to be performed. When the connection is
closed, the error was handled but object was not cleaned. When
performing calls after that point would always append to the internal
buffer. When this is triggered by the cache flush, each seconds a call
would be performed and never stop.

This fixes #26 because memory won't grow on error anymore, but this
doesn't fix the root cause which is a too large inode directory.

There is no more call to hiredis outside zdb.c which centralize all
redis calls to zdb.c handler, there is a new zdb_set_initial function
used to behave like zdb_set but only for id 0 (create or update).

This cleanup and make code more consistant.