directory: add segment indirection list layer
This commit is a major change on how directories are stored and handled
in general. This break all compatibility with any previous version.

== Implementation ==

The first version of directory handling was really simple:
each directories are a single inode and that inode object contains
a list of entries (simple ordered array).

When saving into the backend, the inode is exactly inode header with
each entries concatenated after. This mean that each time the directory
was updated, the full entries list was written again into the database,
which is always append.

New implementation add a new indirection layer to split entries into
multiple objects. Each directory inode contains a new fixed list of
128 lists. Each list will contains part of directory contents.
When adding or removing a file in the directory, only the segment where
that file is, is updated, leaving the inode and others list untouched.

This reduce a lot the amount of data to write into the database on small updates
and reduce the amount of files manipulation when ordering/searching a file.
This reduce array sizes as well.

== Tests ==

Test files are all XX bytes name.

Let's first create a directory with 20k files inside.
Then, 20 times (with 15 seconds delay between each time), add 10 files.

Database usage after the test:
- New implementation: 7.4M (5.8M data, 1.7M index)
- Old implementation: 51M (50M data, 1.7M index)

This is because each change on the directory structure only update
small chunk and doesn't rewrite the full object each time like before.

Note: obviously, if there are lot of changes on each batch which touch
all lists, there won't be difference with old implementation, eveyrthing
will be updated and rewritten.

== Performance ==

There are no real performance improvement or degradation with the new
implementation, at least for now. Performance was not the main goal for
this rewrite.

== Limitations ==

Backend database (zdb) doesn't allow payload larger than 8 MB,
which means an inode won't not be able to be larger than 8 MB.

This was around 80k of our test files.
That was a really hard limit without real workaround.

Using same filenames now put the limit to around 10 millions files
(128 times more) if names are well spread, which should be the case for
an average usage.

A better support of payload size limit should be added.