zdb: better support of namespace mode during autons
When using '-o autons' option, namespaces are created by zdbfs itself.

Namespace have to be in 'sequential' mode and zdbfs attempt to set that
mode via NSSET command when namespace is created. If zdb is running in
--mode seq, an error is triggered when trying to change the mode (even
if it's the good one) because it's running in single mode.

This fix change the workflow of how 'autons' feature:
  For each namespace required (meta, data, temp):
    - First check if the namespace already exists
      - If it doesn't exists, try to create it using NSNEW
    - Check if that namespace is set to sequential mode
      - If it doesn't, try to change the mode using NSSET

If NSNEW fails of NSSET fails, this is fatal and zdbfs stops. Otherwise
everything is fine and no error are reported.

This fixes #43